requirements.txt
requests
ffmpeg-downloader
python-dotenv
Flask
Flask-SQLAlchemy

# WhisperX / transcription deps (merged from whisperx/requirements.txt)
git+https://github.com/openai/whisper.git@main
git+https://github.com/m-bain/whisperx.git@main
ffmpeg-python
numpy==2.0.2
tqdm
torchinfo